feat: create ManagePoolLink component
