+++ apps/main/src/dex/components/PagePool/index.tsx
@@ -24,10 +24,8 @@ import { getChainPoolIdActiveKey } from '@/dex/utils'
 import { getPath } from '@/dex/utils/utilsRouter'
 import { ManageGauge } from '@/dex/widgets/manage-gauge'
 import type { Chain } from '@curvefi/prices-api'
-import Button from '@mui/material/Button'
 import Stack from '@mui/material/Stack'
 import Typography from '@mui/material/Typography'
-import { Link as TanstackLink } from '@tanstack/react-router'
 import { AlertBox } from '@ui/AlertBox'
 import { useUserProfileStore } from '@ui-kit/features/user-profile'
 import { useNavigate } from '@ui-kit/hooks/router'
@@ -40,13 +38,11 @@ import { SizesAndSpaces } from '@ui-kit/themes/design/1_sizes_spaces'
 import { DetailPageLayout } from '@ui-kit/widgets/DetailPageLayout/DetailPageLayout'
 import { FormMargins } from '@ui-kit/widgets/DetailPageLayout/FormTabs'
 import { PoolAlertBanner } from '../PoolAlertBanner'
+import { ManagePoolLink } from './components/ManagePoolLink'
 
 const { Spacing } = SizesAndSpaces
 
 const DEFAULT_SEED: Seed = { isSeed: null, loaded: false }
-
-/** Prices API tells us which pools methods are available, of which the following one is a requisite for refuels */
-const hasRefuelMethod = (poolMethods?: string[]) => poolMethods?.includes('donation_shares')
 
 export const Transfer = (pageTransferProps: PageTransferProps) => {
   const { params, curve, hasDepositAndStake, poolData, poolDataCacheOrApi, routerParams } = pageTransferProps
@@ -85,9 +81,6 @@ export const Transfer = (pageTransferProps: PageTransferProps) => {
   const snapshotData = snapshots?.[0]
 
   const pricesApiPoolData = poolData && pricesApiPoolsMapper?.[poolData.pool.address]
-  const managePoolPath = hasRefuelMethod(pricesApiPoolData?.poolMethods)
-    ? getPath(params, `${ROUTE.PAGE_POOLS}/${poolAddress}/refuel`)
-    : undefined
   const isPoolAdvancedDetailsEnabled = usePoolAdvancedDetails()
 
   type DetailInfoTab = 'user' | 'pool' | 'advanced'
@@ -234,7 +227,6 @@ export const Transfer = (pageTransferProps: PageTransferProps) => {
             poolDataCacheOrApi={poolDataCacheOrApi}
             poolAlert={poolAlert}
             pricesApiPoolData={pricesApiPoolData}
-            managePoolPath={managePoolPath}
           />
         ) : (
           <Stack>
@@ -246,17 +238,7 @@ export const Transfer = (pageTransferProps: PageTransferProps) => {
                 options={poolInfoTabs}
                 testIdPrefix="pool-info-tab"
               />
-              {hasRefuelMethod(pricesApiPoolData?.poolMethods) && (
-                <Button
-                  component={TanstackLink}
-                  to={getPath(params, `${ROUTE.PAGE_POOLS}/${poolAddress}/refuel`)}
-                  variant="inline"
-                  color="ghost"
-                  sx={{ whiteSpace: 'nowrap', alignSelf: 'end', marginBlockEnd: Spacing.xs }}
-                >
-                  {t`Manage pool`}
-                </Button>
-              )}
+              <ManagePoolLink chainId={rChainId} poolAddress={poolAddress} />
             </Stack>
             {poolInfoTab === 'user' && (
               <MySharesStats
